Upgrade django-registration from 0.8 to 1.0.
