+++ requirements/requirements.txt
@@ -5,7 +5,7 @@ pytz==2013b
 django-announcements==1.0.2
 django-celery==3.0.17
 django-timezones==0.2
-django-registration==0.8
+django-registration==1.0
 django_compressor==1.3
 django-email-confirmation==0.2
 django-extensions==1.1.1
